random question and nav
randomized our quiz questions and added navigation to the end of the quiz page.  Tried to work on a new completion page to update the user when the complete the quiz but right now the completed quiz just have links for the user to move around.  quiz_complete.html doesn't have much  use at the moment.